+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## HEAD
 
 - 🚨 **Breaking change:** The default value of `DEPLOY_ENV` now stays `development` regardless of the mode [#74](https://github.com/mapbox/underreact/issues/74).
+- 🚨 **Breaking change:** The default `outputDirectory` is now `_site`. [#71](https://github.com/mapbox/underreact/issues/71)
+